Extract inner PositionsAppender impls to top level classes
